feat: also implement api gateway and sqs functionality
This release adds more functionality to the framework to also process API Gateway and SQS events.
Also adding a mode where we can simply register a provider for a specific type and request it. This will greatly simplify development for e.g. the auth email data and have it requested instead of needing to update it all the time...